requirements.txt
# Core Scientific Computing
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# Data Validation
pydantic>=2.5.0

# Visualization
plotly>=5.18.0
streamlit>=1.29.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.2
pyyaml>=6.0.1
